Fixed constructor in header

+++ src/Team167Lib/SwerveDrive.h
@@ -35,7 +35,7 @@ private:
 
 public:
 	SwerveDrive(int rotateEncLines, int driveEncLines,
-				int feetToEncLinesR,
+				double feetToEncLinesR,
 				short rotateTalon1Number,
 				short rotateTalon2Number,
 				short gyroChannel,
